server module (WIP)
