+++ src/main/scala/krs/service/UserServer.scala
@@ -5,16 +5,16 @@ import com.twitter.finagle.Thrift
 import com.twitter.finagle.stats.Counter
 import com.twitter.server.TwitterServer
 
-import krs.service.impl.{ UserServerImpl }
+import krs.infrastructure.InfrastructureModule
 
-object UserServer extends TwitterServer {
-  val serviceImpl = new UserServerImpl()
+object UserServer extends TwitterServer with ServiceModule {
+  val service = userApi()
   val userService: Counter = statsReceiver.counter("userService")
 
   def main(): Unit = {
     val server = Thrift.server
       .withStatsReceiver(statsReceiver)
-      .serveIface("localhost:8082", serviceImpl())
+      .serveIface("localhost:8082", service)
 
     onExit { server.close() }
     Await.ready(server)
